chore(spp): refresh benchmarks and consolidate test updates
Move external-data hash tests into the explicit integration suite and restore runtime PDA derivation for ring-config creation. Regenerate the CU report from the current program build.
